+++ ss13_tools/scrubby/__init__.py
@@ -1,8 +1,11 @@
-from round_data import RoundData
+# Functions and module names taken directly from Scrubby
+from .round_data import RoundData
+from .CKeyController import GetReceipts
 from .__version__ import __version__
 
 version = __version__
 
 __all__ = [
-    'RoundData'
+    'RoundData',
+    'GetReceipts'
 ]
